Refactor address decoder to 64-bit and implement __builtin_ctz hardware optimization

+++ include/AddressDecoder.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <cstdint>
 
-uint32_t get_index(uint32_t address, uint32_t num_sets, uint32_t block_size);
-uint32_t get_tag(uint32_t address, uint32_t num_sets, uint32_t block_size);
-uint32_t get_offset(uint32_t address, uint32_t block_size);
+uint32_t get_index(uint64_t address, uint32_t num_sets, uint32_t block_size);
+uint64_t get_tag(uint64_t address, uint32_t num_sets, uint32_t block_size);
+uint32_t get_offset(uint64_t address, uint32_t block_size);
